+++ src/com/landawn/abacus/util/FloatTriple.java
@@ -38,7 +38,7 @@ public final class FloatTriple {
     }
 
     public double average() {
-        return sum() / 3d;
+        return (0d + _1 + _2 + _3) / 3;
     }
 
     public FloatTriple reversed() {